separate logic with the dom manipulation
